fix(caddy): update Content-Security-Policy for script-src (#170)
This pull request updates the Content Security Policy (CSP) for the
`photos.stzky.com` site in the `caddy/config/caddy/Caddyfile`. The main
change is to the `script-src` directive, which now allows the use of
`unsafe-eval` in addition to the previously permitted sources.

Security policy update:

* Modified the `Content-Security-Policy` header to include
`'unsafe-eval'` in the `script-src` directive, allowing scripts to use
`eval()` and similar methods. This change may have security
implications, so it's important to ensure that this is necessary for
site functionality. (`caddy/config/caddy/Caddyfile`)